Add files via upload
This is a project that utilizes data held in an excel file which provided to me by a healthcare company as an data science internship project. The data in this case is the results of a survey regarding assault during the COVID-19 pandemic (I did not chose the topic). The programs looks at the answers to 2 questions: 'Do you trust the police?' and 'Did you report assault?'. Then, it answers 2 questions implicitly through analyzing the data: 'How many of those who reported assault, trust the police?' as well as the counter: 'How many of those who trust the police ended up reporting?'. I have included the image file for the output of the program, however I cannot release the excel file publicly as it contains sensitive and personal information of the survey-takers.