Add files via upload
Internship assignment using excel data. This involved a survey pertaining to assault during the COVID-19 pandemic. I chose 2 questions and analyzed the correlation between trusting the police and reporting  assault (I did not chose the topic of the survey). The graphs show the percentage of people who answered favorably to trusting the police and those who answered that they did indeed report assault. Then, of those who do trust the police, did they report (this corollary can be extrapolated to determine if those who trust the police have a propensity to report assault, meaning they are reporting to the police. If they do not trust the police, they would be less inclined to report).